src/main/java/it/sevenbits/handler/inheritors/OpenBracketHandler.java
package it.sevenbits.handler.inheritors;

import it.sevenbits.handler.Handler;
import it.sevenbits.handler.HandlerException;
import it.sevenbits.handler.Indent;
import it.sevenbits.writer.Writable;
import it.sevenbits.writer.WriterException;

/**
 * Handler for open bracket - '{'.
 */
public class OpenBracketHandler extends Handler {

    @Override
    public void handle(final Writable<Character> out, final Indent indent, final char readChar) throws HandlerException {
        try {
            indent.incCurrentIndent();
            out.write(readChar);
            out.write('\n');
            writeIndent(out, (indent.getCurrentIndent() - 1) * indent.getSpaces());
        } catch (WriterException e) {
            throw new HandlerException("error of write in OpenBracketHandler", e);
        }
    }
}
